Revert "Add Command To write install instructrions (#20)"
This reverts commit db32648fad466145dbdbdb0aace1adce170071e3.
