+++ Dockerfile
@@ -5,7 +5,7 @@ COPY goenv /usr/local/bin
 COPY scripts/goenv-test.bash /usr/local/bin/goenv-test
 
 RUN \
-    apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install -y sudo ca-certificates curl && \
+    apt-get update && DEBIAN_FRONTEND=noninteractive apt-get install -y sudo ca-certificates && \
     useradd -m drew && \
     echo "drew ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
