+++ internal/cmd/config.go
@@ -16,10 +16,12 @@ package cmd
 // limitations under the License.
 // ///////////////////////////////////////////////////////////////////////
 import (
+	"bytes"
 	"fmt"
+	"text/tabwriter"
 
-	"github.com/drewgonzales360/goenv/internal/pkg"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 )
 
 type Config struct {
@@ -36,10 +38,50 @@ const (
 
 	GoEnvRootDirEnvVar    = "GOENV_ROOT_DIR"
 	GoEnvInstallDirEnvVar = "GOENV_INSTALL_DIR"
+
+	// Used to let the user know how a config variable was set.
+	configSetByDefault = "\t(default)\n"
+	configSetByEnv     = "\t(set by environment variable)\n"
 )
 
 func ConfigCommand(cmd *cobra.Command, _ []string) {
-	config := pkg.ReadConfig()
+	config := ReadConfig()
 	fmt.Print(config)
-	config.WarnOnMissingPath()
+	warnOnMissingPath(config)
+}
+
+// ReadConfig reads the environment variables for a user and creates a config. If we need any
+// additional config, it'll be parsed in here.
+func ReadConfig() *Config {
+	return &Config{
+		GoenvRootDirectory:    viper.GetString(GoEnvRootDirEnvVar),
+		GoenvInstallDirectory: viper.GetString(GoEnvInstallDirEnvVar),
+	}
+}
+
+// String prints the config to a tabwriter so that the columns are aligned when it's fmt.Print'ed to
+// the terminal.
+func (c *Config) String() string {
+	buf := bytes.Buffer{}
+	w := tabwriter.NewWriter(&buf, 0, 0, 1, ' ', 0)
+
+	rootDir := fmt.Sprintf("%s:\t%s", GoEnvRootDirEnvVar, c.GoenvRootDirectory)
+	if c.GoenvRootDirectory == DefaultGoenvRootDirectory {
+		rootDir += configSetByDefault
+	} else {
+		rootDir += configSetByEnv
+	}
+	w.Write([]byte(rootDir))
+
+	installDir := fmt.Sprintf("%s:\t%s", GoEnvInstallDirEnvVar, c.GoenvInstallDirectory)
+	if c.GoenvInstallDirectory == DefaultGoInstallDirectory {
+		installDir += configSetByDefault
+	} else {
+		installDir += configSetByEnv
+	}
+	w.Write([]byte(installDir))
+
+	w.Flush()
+
+	return buf.String()
 }
