+++ internal/cmd/install.go
@@ -17,20 +17,26 @@ package cmd
 // ///////////////////////////////////////////////////////////////////////
 import (
 	"fmt"
+	"path"
 
 	"github.com/Masterminds/semver/v3"
-	"github.com/drewgonzales360/goenv/internal/pkg"
 	"github.com/spf13/cobra"
+
+	"github.com/drewgonzales360/goenv/internal/pkg"
 )
 
 func InstallCommand(cmd *cobra.Command, args []string) error {
 	version := args[0]
-	config := pkg.ReadConfig()
+	config := ReadConfig()
 
-	return install(config, version)
+	if err := install(config, version); err != nil {
+		return err
+	}
+
+	return nil
 }
 
-func install(config *pkg.Config, version string) error {
+func install(config *Config, version string) error {
 	if inaccessible := pkg.CheckRW(config.GoenvRootDirectory, config.GoenvInstallDirectory); len(inaccessible) > 0 {
 		return fmt.Errorf(PermError, inaccessible)
 	}
@@ -45,7 +51,7 @@ func install(config *pkg.Config, version string) error {
 		return err
 	}
 
-	err = pkg.ExtractTarGz(tarballPath, config.VersionInstallPath(goVersion))
+	err = pkg.ExtractTarGz(tarballPath, path.Join(config.GoenvInstallDirectory, goVersion.String()))
 	if err != nil {
 		return fmt.Errorf("could not extract go: %w", err)
 	}
